refactor(wiremock): switches from config resolver to plugin.
