+++ core/config-resolver-wiremock/build.gradle
@@ -8,12 +8,13 @@ compileJava {
 }
 
 dependencies {
-    implementation project(':core:imposter-api')
-    implementation project(':core:config')
+    implementation project(':core:imposter-engine')
     implementation project(':mock:mock-rest')
 
     // test
     testImplementation project(':test:test-utils')
+    testImplementation group: 'org.hamcrest', name: 'hamcrest', version: version_hamcrest
+    testImplementation "org.mockito.kotlin:mockito-kotlin:4.0.0"
 
     // java 11 compatibility
     testImplementation ("javax.xml.bind:jaxb-api:$version_jaxb_api")
