--- core/config-resolver-wiremock/src/main/resources/META-INF/config-resolver.properties
@@ -1,1 +1,0 @@
-class=io.gatehill.imposter.plugin.wiremock.config.WiremockConfigResolver
