--- core/config-resolver-wiremock/src/test/resources/config/imposter-config.yaml
@@ -1,3 +1,0 @@
-plugin: "rest"
-response:
-  content: "Hello world"
