+++ core/config-resolver-wiremock/src/test/resources/wiremock-simple/imposter-config.yaml
@@ -1,0 +1,1 @@
+plugin: wiremock
