+++ core/config-resolver-wiremock/src/test/resources/wiremock-xpath-request/mappings/mappings/mappings.json
